only delete the secret on test and prod

+++ app/controllers/recipients/secrets_controller.rb
@@ -11,7 +11,7 @@ module Recipients
 
     def create
       if secret.decrypt
-        secret.destroy
+        secret.destroy unless Rails.env.development?
         render :create
       else
         render :new
